+++ react/IconAssistantSales.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconAssistantSales = (props: IconProps) => {
+  return <Icon id="hpa-telemarketing" {...props} />
+}
+
+export default IconAssistantSales