+++ react/IconCart.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconCart = (props: IconProps) => {
+  return <Icon id="hpa-cart" {...props} />
+}
+
+export default IconCart