+++ react/IconDelete.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconDelete = (props: IconProps) => {
+  return <Icon id="hpa-delete" {...props} />
+}
+
+export default IconDelete