+++ react/IconEquals.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconEquals = (props: IconProps) => {
+  return <Icon id="sti-equals" {...props} />
+}
+
+export default IconEquals