+++ react/IconEyeSight.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+import { getType, getState } from './utils/helpers'
+
+const IconEyeSight = ({ type, state, ...props }: EnhancedIconProps) => {
+  const typeModifier = getType(type, 'filled, outline')
+  const stateModifier = getState(state)
+  return <Icon id={`mpa-eyesight${typeModifier}${stateModifier}`} {...props} />
+}
+
+export default IconEyeSight