+++ react/IconFilter.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconFilter = (props: IconProps) => {
+  return <Icon id="mpa-filter-settings" {...props} />
+}
+
+export default IconFilter