+++ react/IconGrid.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconGrid = (props: IconProps) => {
+  return <Icon id="mpa-gallery" {...props} />
+}
+
+export default IconGrid