+++ react/IconHome.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconHome = (props: IconProps) => {
+  return <Icon id="nav-home" {...props} />
+}
+
+export default IconHome