+++ react/IconInlineGrid.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconInlineGrid = (props: IconProps) => {
+  return <Icon id="mpa-list-items" {...props} />
+}
+
+export default IconInlineGrid