+++ react/IconPlus.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconPlus = (props: IconProps) => {
+  return <Icon id="nav-plus" {...props} />
+}
+
+export default IconPlus