+++ react/IconProfile.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconProfile = (props: IconProps) => {
+  return <Icon id="hpa-profile" {...props} />
+}
+
+export default IconProfile