+++ react/IconRemove.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconRemove = (props: IconProps) => {
+  return <Icon id="mpa-remove" {...props} />
+}
+
+export default IconRemove