+++ react/IconSingleGrid.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconSingleGrid = (props: IconProps) => {
+  return <Icon id="mpa-single-item" {...props} />
+}
+
+export default IconSingleGrid