+++ react/IconSwap.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+import Icon from './components/Icon'
+
+const IconSwap = (props: IconProps) => {
+  return <Icon id="mpa-swap" {...props} />
+}
+
+export default IconSwap